object-detection/requirements.txt
tensorflow-gpu==2.4.1
opencv-python==4.2.0.34
keras==2.4.3 
pillow==7.0.0 
scipy==1.4.1 
h5py==2.10.0
imageai
